Primary and foreign keys are specified.
NOTE: After implementing foreign key not-null constraints, the delete
functions would need to get a more complex logic. However, as those are not
needed for the new database model to function, this is not done. Therefore
only the last two most important functions are used to demonstrate the working
database model.